requirements.txt.txt
fastapi
uvicorn[standard]
httpx
pydantic
better_profanity
dnspython